+++ src/main/java/umc/study/converter/TempConverter.java
@@ -1,6 +1,6 @@
 package umc.study.converter;
 
-import umc.study.web.dto.TempResponse;
+import umc.study.web.dto.Temp.TempResponse;
 
 public class TempConverter {
