+++ src/main/java/umc/study/repository/MissionRepository/MissionRepositoryCustom.java
@@ -3,8 +3,8 @@ package umc.study.repository.MissionRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import umc.study.domain.enums.MissionStatus;
-import umc.study.web.dto.MemberMissionDto;
-import umc.study.web.dto.RegionMissionDto;
+import umc.study.web.dto.Member.MemberMissionDto;
+import umc.study.web.dto.Region.RegionMissionDto;
 
 public interface MissionRepositoryCustom {
     public Page<MemberMissionDto> findByMemberAndStatus(Long memberId, MissionStatus status, Pageable pageable);
