+++ src/main/java/umc/study/service/MemberService/MemberCommandService.java
@@ -1,7 +1,7 @@
 package umc.study.service.MemberService;
 
 import umc.study.domain.Member;
-import umc.study.web.dto.MemberRequestDTO;
+import umc.study.web.dto.Member.MemberRequestDTO;
 
 public interface MemberCommandService {
