+++ src/main/java/umc/study/service/MemberService/MemberCommandServiceImpl.java
@@ -12,7 +12,7 @@ import umc.study.domain.Member;
 import umc.study.domain.mapping.MemberPrefer;
 import umc.study.repository.FoodCategoryRepository;
 import umc.study.repository.MemberRepository.MemberRepository;
-import umc.study.web.dto.MemberRequestDTO;
+import umc.study.web.dto.Member.MemberRequestDTO;
 
 import java.util.List;
 import java.util.stream.Collectors;
