+++ src/main/java/umc/study/web/controller/MemberRestController.java
@@ -11,8 +11,8 @@ import umc.study.apiPayload.ApiResponse;
 import umc.study.converter.MemberConverter;
 import umc.study.domain.Member;
 import umc.study.service.MemberService.MemberCommandService;
-import umc.study.web.dto.MemberRequestDTO;
-import umc.study.web.dto.MemberResponseDTO;
+import umc.study.web.dto.Member.MemberRequestDTO;
+import umc.study.web.dto.Member.MemberResponseDTO;
 
 @RestController
 @RequiredArgsConstructor
