+++ src/main/java/umc/study/web/dto/Member/MemberResponseDTO.java
@@ -1,4 +1,4 @@
-package umc.study.web.dto;
+package umc.study.web.dto.Member;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
