+++ src/main/java/umc/study/web/dto/Region/RegionMissionDto.java
@@ -1,13 +1,12 @@
-package umc.study.web.dto;
+package umc.study.web.dto.Region;
+
 
 import java.time.LocalDate;
 
-public record MemberMissionDto(
+public record RegionMissionDto(
         Long missionId,
         String missionSpec,
         Integer reward,
         LocalDate deadline,
-        String storeName,
-        String status
-) {
-}
+        String storeName
+) {}
