+++ src/main/java/umc/study/web/dto/Temp/TempRequest.java
@@ -1,0 +1,5 @@
+package umc.study.web.dto.Temp;
+
+public class TempRequest
+{
+}
