+++ src/main/java/umc/study/web/dto/Temp/TempResponse.java
@@ -1,4 +1,4 @@
-package umc.study.web.dto;
+package umc.study.web.dto.Temp;
 
 
 import lombok.AllArgsConstructor;
