--- src/main/java/umc/study/web/dto/TempRequest.java
@@ -1,5 +1,0 @@
-package umc.study.web.dto;
-
-public class TempRequest
-{
-}
